(feat) Hub from our own minimal image to reduce vulnrabilities, have more control, and reduce size
Could split into more layers, but it doesn't take that long to build (1-2
mins), and going for as small an image as possible.

Deliberatly pinning as much as possible for (close-to) reproducable builds.

The PATH environment variable is not overridden in the notebook, since if we do
the notebook doesn't start. The notebook can just manage its own path. No need
for the hub and the notebook to be coupled together in this way. Suspect it is
only good for process-based local notebooks.